+++ src/services/TrackerService.ts
@@ -1,4 +1,4 @@
-import { createCatchHandler, createSuccessHandler } from '@/composables/service-helpers';
+import { createCatchHandler } from '@/composables/service-helpers';
 
 import { osidbFetch } from '@/services/OsidbAuthService';
 import { osimRuntime } from '@/stores/osimRuntime';
@@ -16,7 +16,6 @@ export async function fileTrackingFor(trackerData: TrackersPost | TrackersPost[]
   if (!Array.isArray(trackerData) || trackerData.length === 1) {
     const tracker = !Array.isArray(trackerData) ? trackerData : trackerData[0];
     return postTracker(tracker)
-      .then(createSuccessHandler({ title: 'Success!', body: 'Tracker filed.', css: 'success' }))
       .catch(createCatchHandler(`Failed to create tracker for ${tracker.ps_update_stream}`));
   }
 
@@ -51,11 +50,6 @@ export async function fileTrackingFor(trackerData: TrackersPost | TrackersPost[]
     createCatchHandler(`${errors.length} trackers failed to file`)(messages);
     return Promise.reject({ errors, successes });
   } else {
-    createSuccessHandler({
-      title: 'Success!',
-      body: `${trackerData.length} trackers filed.`,
-      css: 'success',
-    })({ data: null });
     return Promise.resolve({ successes });
   }
 }